Rewrite InitCommand.
